Exceptions on Invalid Orders
Created exceptions to better debug the invalid orders.
Created a variable to define the end year of the game.
Created a variable to set the mode of the game. This enables the negotiation server or disables it.
Tried to figure out why there are some orders that are generated on invalid provinces but still didnt reach a conclusion. Maybe it is on the hold orders that are created to prevent bad support orders. TODO: Solve the invalid orders.